Extract local AI provider controls
